+++ tests/testthat/test-glm.R
@@ -93,18 +93,18 @@ withr::with_environment(
       data = lme4::cbpp
     )
     expect_s3_class(
-      ggpredict(fit, "c12hour", vcov_fun = "vcovHC", vcov_type = "HC1", verbose = FALSE),
+      ggpredict(fit, "c12hour", vcov = "HC1", verbose = FALSE),
       "data.frame"
     )
     expect_s3_class(
-      ggpredict(fit, c("c12hour", "c161sex"), vcov_fun = "vcovHC", vcov_type = "HC1", verbose = FALSE),
+      ggpredict(fit, c("c12hour", "c161sex"), vcov = "HC1", verbose = FALSE),
       "data.frame"
     )
     expect_s3_class(
-      ggpredict(fit, c("c12hour", "c161sex", "c172code"), vcov_fun = "vcovHC", vcov_type = "HC1", verbose = FALSE),
+      ggpredict(fit, c("c12hour", "c161sex", "c172code"), vcov = "HC1", verbose = FALSE),
       "data.frame"
     )
-    expect_s3_class(ggpredict(m, "period", vcov_fun = "vcovHC", vcov_type = "HC1"), "data.frame")
+    expect_s3_class(ggpredict(m, "period", vcov = "HC1"), "data.frame")
   })
 )